backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.23
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.9
requests==2.32.3
cloudinary==1.41.0
email-validator==2.2.0
faker==24.0.0
slowapi==0.1.9